requirements.txt
ollama==0.5.3
pandas==2.3.1
tqdm==4.67.1
fast-bleu==0.0.90
huggingface-hub==0.34.4
pyarrow==21.0.0
unbabel-comet==2.2.6
python-dotenv==1.1.1
bert_score==0.3.13